Refactor to split utils file
