+++ main.py
@@ -2,7 +2,8 @@ import asyncio
 from kasa import Discover
 from dotenv import load_dotenv
 import os
-from strom import utils
+from strom import optimization_utils
+from strom.data_utils import get_temp_price_df
 
 # Load the environment variables from the .env file
 load_dotenv(dotenv_path="../../config/tapologin.env")
@@ -11,7 +12,7 @@ email = os.getenv("EMAIL")  # Get email from the environment variable
 password = os.getenv("PASSWORD")  # Get password from the environment variable
 device_ip = os.getenv("DEVICEIP")
 if not device_ip:
-    os.chdir(utils.find_root_dir())
+    os.chdir(optimization_utils.find_root_dir())
     with open('./config/device_IP.txt') as f:
         device_ip = f.read().strip()
 
@@ -19,9 +20,9 @@ async def main():
     try:
         # Discover the device
         dev = await Discover.discover_single(device_ip, username=email, password=password)
-        temp_price_df = utils.get_temp_price_df()
+        temp_price_df = get_temp_price_df()
         # Prompt the user for input (0 for off, 1 for on)
-        user_input = bool(utils.find_heating_decision(temp_price_df, decision = 'discrete')[0][0])
+        user_input = bool(optimization_utils.find_heating_decision(temp_price_df, decision = 'discrete')[0][0])
         # Check user input and turn the light on or off accordingly
         print(user_input)
         if user_input:
